// src/pages/manageTeam/types.ts
export type DepartmentStats = {
    name: string;
    point: number;
    leader: string;
    dispatch: number;
    assistant: number;
    driver: number;
    score: number;
};

export type Dispatch = {
    name: string;
    age: number
}

export type Assistant = {
    name: string;
}

export type Driver = {
    name: string;
}